+++ lib/rubocop/cop/style/stabby_lambda_parentheses.rb
@@ -25,8 +25,6 @@ module RuboCop
         MSG_REQUIRE = 'Wrap stabby lambda arguments with parentheses.'.freeze
         MSG_NO_REQUIRE = 'Do not wrap stabby lambda arguments ' \
                          'with parentheses.'.freeze
-        ARROW = '->'.freeze
-
         def on_send(node)
           return unless stabby_lambda_with_args?(node)
           return unless redundant_parentheses?(node) ||
@@ -76,11 +74,7 @@ module RuboCop
         end
 
         def stabby_lambda_with_args?(node)
-          node.command?(:lambda) && arrow_form?(node) && node.parent.arguments?
-        end
-
-        def arrow_form?(node)
-          node.loc.selector.source == ARROW
+          node.stabby_lambda? && node.parent.arguments?
         end
 
         def parentheses?(node)